refactor: 패키지 구조 도메인 별로 변경, 도메인 별 Dao 추상 클래스 정의
- 기본적인 CRUD 관련 메서드 표준을 제공하는 Dao 인터페이스를 두되
- 도메인 별로 사용하지 않는 메서드/추가할 메서드를 정의하기 위해 추상 클래스 정의
- 추상 클래스의 접근 제어자로 의도치 않은 상속 방지하기 위해 패키지 구조 변경함.
- 학습 목적으로 진행하였으나 각 도메인 별 Dao의 인터페이스를 다르게 가지는 것이 낫다고 판단되면 다시 수정 예정.
