+++ src/main/java/cart/domain/member/Email.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.member;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @EqualsAndHashCode
-public class Password {
+public class Email {
 
     private final String value;
 }
