+++ src/main/java/cart/domain/member/H2MemberDao.java
@@ -1,8 +1,5 @@
-package cart.dao;
+package cart.domain.member;
 
-import cart.domain.Email;
-import cart.domain.Member;
-import cart.domain.Password;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -10,7 +7,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class H2MemberDao implements Dao<Member> {
+public class H2MemberDao extends MemberDao {
 
     private static final RowMapper<Member> MEMBER_ROW_MAPPER = (resultSet, rowNum) -> new Member(
             resultSet.getLong("id"),
@@ -24,32 +21,12 @@ public class H2MemberDao implements Dao<Member> {
     }
 
     @Override
-    public Member insert(final Member entity) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void update(final Member entity) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean isExist(final Long id) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Optional<Member> findById(final Long id) {
-        return Optional.empty();
-    }
-
-    @Override
     public List<Member> findAll() {
         return jdbcTemplate.query("SELECT id, email, password FROM members", MEMBER_ROW_MAPPER);
     }
 
     @Override
-    public void deleteById(final Long id) {
-        throw new UnsupportedOperationException();
+    Optional<Member> findByEmail(final Email email) {
+        return Optional.empty();
     }
 }
