+++ src/main/java/cart/domain/product/ProductDao.java
@@ -1,0 +1,6 @@
+package cart.domain.product;
+
+import cart.dao.Dao;
+
+abstract class ProductDao implements Dao<Product> {
+}
