+++ src/main/java/cart/domain/product/ProductService.java
@@ -1,16 +1,14 @@
-package cart.service;
+package cart.domain.product;
 
-import cart.dao.Dao;
-import cart.domain.Product;
 import java.util.List;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ProductService {
 
-    private final Dao<Product> productDao;
+    private final ProductDao productDao;
 
-    public ProductService(final Dao<Product> productDao) {
+    public ProductService(final ProductDao productDao) {
         this.productDao = productDao;
     }
